+++ management/model_enum_certificate_key_algorithm.go
@@ -15,7 +15,7 @@ import (
 	"fmt"
 )
 
-// EnumCertificateKeyAlgorithm Specifies the key algorithm. Options are RSA, EC, and UNKNOWN.
+// EnumCertificateKeyAlgorithm Specifies the key algorithm. Options are `RSA`, `EC`, and `UNKNOWN`.
 type EnumCertificateKeyAlgorithm string
 
 // List of EnumCertificateKeyAlgorithm